+++ packages/miniapp-webview-config/src/index.ts
@@ -13,15 +13,22 @@ export function setWebviewConfig(config, options) {
     getValue,
     context: {
       userConfig: rootUserConfig,
-      rootDir
+      rootDir,
+      webpack
     },
-    cancelTask
+    applyMethod,
+    hasMethod,
+    cancelTask,
+    log
   } = api;
+
+  const isWebpack4 = /^4\./.test(webpack.version);
   const userConfig = rootUserConfig[target] || {};
 
   // If using frm then do not generate miniapp webview code temporarily
   if (target === MINIAPP && userConfig.frm === true ) {
     cancelTask(MINIAPP);
+    return;
   }
 
   const appConfig = normalizeStaticConfig(getValue('staticConfig'), { rootDir });
@@ -58,6 +65,36 @@ export function setWebviewConfig(config, options) {
       nativeConfig: userConfig.nativeConfig,
     },
   ]);
+
+  config
+    .output
+    .library(['page', '[name]'])
+    .libraryTarget('umd')
+    .globalObject('self={}');;
+
+  if (!isWebpack4) {
+    config.merge({
+      devServer: {
+        client: false,
+      }
+    })
+  } else {
+    config.devServer.inline(false);
+  }
+
+  if (command === 'start' && config.get('devtool')) {
+    config.devtool('inline-source-map');
+  }
+
+  config.devServer.hot(false);
+  
+  injectJSSDK(hasMethod, applyMethod, target);
+  applyMethod('addPluginTemplate', join(__dirname, './runtime/page.js'));
+  const importDeclarations = getValue('importDeclarations');
+  importDeclarations.createWebviewPage = {
+    value: '$$framework/plugins/miniapp/page'
+  };
+  api.setValue('importDeclarations', importDeclarations);
 };
 
 export function setWebviewPageConfig(config, options, remoteRoutes: RouteType[]) {